+++ gladius_shared/src/error.rs
@@ -39,8 +39,11 @@ pub enum SlicerErrors {
     ///Error during tower generation
     TowerGeneration,
 
-    ///No inout models provided
+    ///No input models provided
     NoInputProvided,
+
+    ///All input must be UTF8
+    InputNotUTF8,
 
     ///Input string is misformated
     InputMisformat,
@@ -53,6 +56,21 @@ pub enum SlicerErrors {
 
     ///Error during tower generation
     SliceGeneration,
+
+    ///File permission issue will settings file or folder
+    SettingsFilePermission,
+
+    ///Failed to create new file
+    FileCreateError {
+        ///File that was not created
+        filepath: String,
+    },
+
+    ///Failed to write to file
+    FileWriteError {
+        ///File that was no able to write to
+        filepath: String,
+    },
 
     ///Error because settings less than zero
     SettingLessThanZero {
@@ -80,6 +98,9 @@ impl SlicerErrors {
     ///Return the error code and pretty error message
     pub fn get_code_and_message(&self) -> (u32, String) {
         match self {
+            SlicerErrors::UnspecifiedError(err_string) => {
+                (0xFFFFFFFF,format!("Third Party Error: {}.",err_string))
+            }
             SlicerErrors::ObjectFileNotFound { filepath } => {
                 (0x1000,format!("Could not load object file \"{}\". It was not found in the filesystem. Please check that the file exists and retry.",filepath))
             }
@@ -122,10 +143,18 @@ impl SlicerErrors {
             SlicerErrors::SettingLessThanOrEqualToZero { setting, value } => {
                 (0x100D,format!("The setting {} must be greater than to 0. It is currently {}.",setting, value))
             }
-            SlicerErrors::UnspecifiedError(err_string) => {
-                (0xFFFFFFFF,format!("Third Party Error. {}",err_string))
+            SlicerErrors::InputNotUTF8 => {
+                (0x100E,"Input String must be UTF8.".to_string())
             }
-
+            SlicerErrors::SettingsFilePermission => {
+                (0x100F,"File permission issue will settings file or folder.".to_string())
+            }
+            SlicerErrors::FileCreateError { filepath } => {
+                (0x1010,format!("Could not create file \"{}\".",filepath))
+            }
+            SlicerErrors::FileWriteError { filepath } => {
+                (0x1011,format!("Could not write to file \"{}\".",filepath))
+            }
         }
     }
 }