+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Fixes
 - Fixed issue when slicing multiple objects and outputting bincode overlapping and conflicting 
+- Removed all unwraps. Converted some to errors and some to expect. deny(clippy::unwrap_used) is added to prevent unwrap being added in future
 
 ## [0.3.0]
 ### New Features